+++ templates/pwa/serviceworker_js.tmpl
@@ -7,6 +7,7 @@ var urlsToCache = [
   '{{StaticUrlPrefix}}/vendor/plugins/fomantic/semantic.min.js?v={{MD5 AppVer}}',
   '{{StaticUrlPrefix}}/js/index.js?v={{MD5 AppVer}}',
   '{{StaticUrlPrefix}}/js/swagger.js?v={{MD5 AppVer}}',
+  '{{StaticUrlPrefix}}/js/contextPopup.js?v={{MD5 AppVer}}',
   '{{StaticUrlPrefix}}/js/gitgraph.js?v={{MD5 AppVer}}',
   '{{StaticUrlPrefix}}/vendor/plugins/clipboard/clipboard.min.js',
   '{{StaticUrlPrefix}}/vendor/plugins/vue/vue.min.js',
